Modifications for PyPi
