+++ src/inso/__init__.py
@@ -1,0 +1,2 @@
+
+from .figures import display
