+++ src/inso/astrofiles/Laskar1993/INSOLN.LA93_01.BTL.ASC
